feat(pluginPreloader): remove measuring and filtering logic
The measuring of the whole preload process won't be necessary
anymore (or at least not here) since we will load app plugins in a "lazy" manner.

The filtering logic is not necessary either, as we will preloading plugins
regardless of the preloading=true flag when it is necessary, and not during the initial
page load. Plugins that has preloading as a hard requirement can be added as exceptions.